RavenDB-17771 initial implementation of sharded changes API
