+++ src/Raven.Server/Documents/Changes/SendQueueItem.cs
@@ -1,0 +1,7 @@
+﻿namespace Raven.Server.Documents.Changes;
+
+public class SendQueueItem
+{
+    public object ValueToSend;
+    public bool AllowSkip;
+}
