+++ src/Raven.Server/Documents/DocumentsTransaction.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Raven.Client.Documents.Changes;
+using Raven.Server.Documents.Changes;
 using Raven.Server.Documents.Replication;
 using Raven.Server.Documents.Replication.Incoming;
 using Raven.Server.ServerWide;
