+++ src/Raven.Server/Documents/Operations/Operations.cs
@@ -11,6 +11,7 @@ using Raven.Client.Exceptions;
 using Raven.Client.Exceptions.Documents;
 using Raven.Client.Extensions;
 using Raven.Client.Util;
+using Raven.Server.Documents.Changes;
 using Raven.Server.NotificationCenter.Notifications;
 using Raven.Server.ServerWide;
 using Raven.Server.Utils;
