+++ src/Raven.Server/Documents/Subscriptions/SubscriptionConnectionsStateForShard.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using Raven.Server.Documents.Sharding;
 using Raven.Server.Documents.TcpHandlers;
 using Raven.Server.ServerWide.Commands.Subscriptions;
