+++ test/FastTests/Client/Changes/BasicChangesTests.cs
@@ -1,0 +1,83 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Tests.Infrastructure;
+using Tests.Infrastructure.Entities;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace FastTests.Client.Changes;
+
+public class BasicChangesTests : RavenTestBase
+{
+    public BasicChangesTests(ITestOutputHelper output) : base(output)
+    {
+    }
+
+    [RavenTheory(RavenTestCategory.ChangesApi)]
+    [RavenData(DatabaseMode = RavenDatabaseMode.All)]
+    public async Task Can_Subscribe_To_Single_Document_Changes(Options options)
+    {
+        using (var store = GetDocumentStore(options))
+        {
+            using (var changes = store.Changes())
+            {
+                await changes.EnsureConnectedNow();
+
+                const int numberOfDocuments = 10;
+
+                var cde = new CountdownEvent(numberOfDocuments);
+
+                for (var i = 0; i < numberOfDocuments; i++)
+                {
+                    changes
+                        .ForDocument($"orders/{i}")
+                        .Subscribe(x => cde.Signal());
+                }
+
+                for (var i = 0; i < numberOfDocuments; i++)
+                {
+                    using (var session = store.OpenAsyncSession())
+                    {
+                        await session.StoreAsync(new Order(), $"orders/{i}");
+                        await session.SaveChangesAsync();
+                    }
+                }
+
+                Assert.True(cde.Wait(TimeSpan.FromSeconds(30)), $"Missed {cde.CurrentCount} events.");
+            }
+        }
+    }
+
+    [RavenTheory(RavenTestCategory.ChangesApi)]
+    [RavenData(DatabaseMode = RavenDatabaseMode.All)]
+    public async Task Can_Subscribe_To_All_Document_Changes(Options options)
+    {
+        using (var store = GetDocumentStore(options))
+        {
+            using (var changes = store.Changes())
+            {
+                await changes.EnsureConnectedNow();
+
+                const int numberOfDocuments = 10;
+
+                var cde = new CountdownEvent(numberOfDocuments);
+
+                changes
+                    .ForAllDocuments()
+                    .Subscribe(x => cde.Signal());
+
+                for (var i = 0; i < numberOfDocuments; i++)
+                {
+                    using (var session = store.OpenAsyncSession())
+                    {
+                        await session.StoreAsync(new Order(), $"orders/{i}");
+                        await session.SaveChangesAsync();
+                    }
+                }
+
+                Assert.True(cde.Wait(TimeSpan.FromSeconds(30)), $"Missed {cde.CurrentCount} events.");
+            }
+        }
+    }
+}
